Allow adding multiple subsections, create `notepad` module, remove `Notepad` from the top-level API, improve tests.
